dashboard/pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "integration-hub-dashboard"
version = "0.1.0"
description = "NHS Wales Integration Hub NOC monitoring dashboard"
requires-python = ">=3.13"

dependencies = [
    "azure-identity>=1.25.0",
    "azure-mgmt-appcontainers>=4.0.0",
    "azure-mgmt-servicebus>=9.0.0",
    "azure-monitor-query>=2.0.0",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
]

[dependency-groups]
dev = [
    "ruff==0.14.9",
    "bandit==1.9.2",
    "mypy==1.18.2",
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
]

[tool.mypy]
python_version = "3.13"
disallow_untyped_defs = true

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # Pycodestyle errors
    "F",    # Pyflakes errors
    "W",    # Pycodestyle warnings
    "A",    # Flake8-builtins
    "PLC",  # Pylint convention
    "PLE",  # Pylint errors
    "PLW",  # Pylint warnings
    "I",    # Import sorting
]

[tool.bandit.assert_used]
skips = ["*/test_*.py", "*/*_test.py"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["dashboard*"]

[tool.setuptools.package-data]
dashboard = ["templates/*.html", "static/**/*"]
